+++ DataAccessInterfaces/IEdubaseRepository.cs
@@ -6,10 +6,10 @@ namespace SFB.Web.ApplicationCore.DataAccess
 {
     public interface IEdubaseRepository
     {
-        Task<EdubaseDataObject> GetSchoolDataObjectByUrnAsync(int urn);
+        Task<EdubaseDataObject> GetSchoolDataObjectByUrnAsync(long urn);
         Task<List<EdubaseDataObject>> GetSchoolsByLaEstabAsync(string laEstab, bool openOnly);        
-        Task<List<EdubaseDataObject>> GetMultipleSchoolDataObjectsByUrnsAsync(List<int> urns);
-        Task<List<int>> GetAllSchoolUrnsAsync();
+        Task<List<EdubaseDataObject>> GetMultipleSchoolDataObjectsByUrnsAsync(List<long> urns);
+        Task<List<long>> GetAllSchoolUrnsAsync();
         Task<IEnumerable<EdubaseDataObject>> GetAcademiesByCompanyNoAsync(int companyNo);
         Task<int> GetAcademiesCountByCompanyNoAsync(int companyNo);
         Task<IEnumerable<EdubaseDataObject>> GetAcademiesByUidAsync(int uid);
