+++ Entities/AcademySummaryDataObject.cs
@@ -6,7 +6,7 @@ namespace SFB.Web.ApplicationCore.Entities
     public class AcademySummaryDataObject
     {
         [JsonProperty(PropertyName = SchoolTrustFinanceDataFieldNames.URN)]
-        public int URN;
+        public long URN;
 
         [JsonProperty(PropertyName = SchoolTrustFinanceDataFieldNames.ESTAB_NAME)]
         public string EstablishmentName;
