+++ Services/DataAccess/IContextDataService.cs
@@ -6,10 +6,10 @@ namespace SFB.Web.ApplicationCore.Services.DataAccess
 {
     public interface IContextDataService
     {
-        Task<EdubaseDataObject> GetSchoolDataObjectByUrnAsync(int urn);        
-        Task<List<int>> GetAllSchoolUrnsAsync();
+        Task<EdubaseDataObject> GetSchoolDataObjectByUrnAsync(long urn);        
+        Task<List<long>> GetAllSchoolUrnsAsync();
         Task<List<EdubaseDataObject>> GetSchoolDataObjectByLaEstabAsync(string laEstab, bool openOnly);        
-        Task<List<EdubaseDataObject>> GetMultipleSchoolDataObjectsByUrnsAsync(List<int> urns);
+        Task<List<EdubaseDataObject>> GetMultipleSchoolDataObjectsByUrnsAsync(List<long> urns);
         Task<IEnumerable<EdubaseDataObject>> GetAcademiesByCompanyNumberAsync(int companyNo);
         Task<IEnumerable<EdubaseDataObject>> GetAcademiesByUidAsync(int uid);
         Task<int> GetAcademiesCountByCompanyNumberAsync(int companyNo);
